The PSX function finder now exports in markdown
